% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLdiv.R
\name{KLdivergence}
\alias{KLdivergence}
\title{Kullback Leibler divergence 
Calculates the KL divergence between the random cluster model and the erdos-renyi model}
\usage{
KLdivergence(x, n.nodes, p.true, p.mod)
}
\arguments{
\item{x}{Matrix with sampled configurations}

\item{n.nodes}{The number of nodes in the model}

\item{p.true}{The p vthat was sampled with in the erdos renyi model}

\item{p.mod}{The p value of the model}
}
\value{
The kl divergence between the Erdos-Renyi model and the Random Cluster model
}
\description{
Kullback Leibler divergence 
Calculates the KL divergence between the random cluster model and the erdos-renyi model
}
