% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{RandomnessRecycler}
\alias{RandomnessRecycler}
\title{Randomness Recycler}
\usage{
RandomnessRecycler(n.nodes, p, q)
}
\arguments{
\item{n.nodes}{The number of nodes in the network}

\item{p}{The wiring probability}

\item{q}{The clustering coefficient of the Random Cluster model}
}
\value{
A adjacency matrix of a graph sampled from the Random Cluster model.
}
\description{
Randomness recycler method for sampling from the Random Cluster model. Introduced by
Fill and Huber (2000).
}
