% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCProb.R
\name{RCProb}
\alias{RCProb}
\title{Random Cluster Probability}
\usage{
RCProb(conf, p, q, n.nodes, Curie = F)
}
\arguments{
\item{conf}{A binary vector with the edge configuration}

\item{p}{Probability of forming an edge between two nodes}

\item{q}{Clustering coefficient RC model}

\item{n.nodes}{The number of nodes in the network}

\item{Curie}{Whether to use the Curie-Weiss partition function, is only possible when q = 2.}
}
\value{
Returns the configuration probability.
}
\description{
Calculates the probability of a edge configuration given parameters p and q.
}
