% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleRC.R
\name{SampleRC}
\alias{SampleRC}
\title{Sample Random Cluster}
\usage{
SampleRC(p, q, n.edges, n.nodes, n.iter, verbose = TRUE)
}
\arguments{
\item{p}{The probability of forming an edge}

\item{q}{The cluster coefficient from the Random Cluster model}

\item{n.edges}{The number of edges in the model}

\item{n.nodes}{The number of nodes in the network}

\item{n.iter}{The maximum number of iterations to run CFTP}

\item{verbose}{Whether to print the progress.}
}
\value{
Returns a matrix with samples from the random cluster model.
}
\description{
Function to sample from the Random Cluster model. Currently this function works with 
the coupling from the past algorithm that is used in Grimmit (2006).
}
